requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
SQLAlchemy==2.0.31
opencv-python==4.10.0.84
Pillow==10.4.0
numpy>=1.22,<1.24
deepface==0.0.93
tensorflow==2.12.0
keras==2.12.0
mtcnn==0.1.1
Flask-Migrate==4.0.7
